+++ src/App.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import './App.css';
+import Rockets from './pages/Rockets';
+import Missions from './pages/Missions';
+import MyProfile from './pages/MyProfile';
+import MyNavbar from './components/Navbar';
+
+function App() {
+  return (
+    <BrowserRouter>
+      <MyNavbar />
+      <Routes>
+        <Route exact path="/" element={<Rockets />} />
+        <Route exact path="/missions" element={<Missions />} />
+        <Route exact path="/myprofile" element={<MyProfile />} />
+      </Routes>
+    </BrowserRouter>
+  );
+}
+
+export default App;
